+++ test/mocks/MockPrompts.ts
@@ -1,0 +1,9 @@
+import { PromptObject, PromptType } from 'prompts';
+
+const prompts = jest.fn();
+
+export {
+  PromptObject,
+  PromptType
+};
+export default prompts;